Rename "Common" directory to "tote_bag"
I wasn't sure where to put this. It's not a "library",
per se, so why put it in libs?

Well, I don't think tote_bag code should go under source,
since that directory is meant to be Valentine specific.

If this repo were meant to accommodate more than one project,
we could do something like putting the common code under source
and valentine code in a dir called "valentine" inside of source.

Furthermore, tote_bag has code that I may want to use on other
projects. In that case, it should be built as an actual library
and included as a submodule.
