+++ Source/GUI/Panels/ValentineCenterPanel.cpp
@@ -13,8 +13,8 @@
 #include "PluginProcessor.h"
 #include "ValentineParameters.h"
 
-#include "Common/GUI/LookAndFeel/LookAndFeel.h"
-#include "Common/GUI/Utilities/GraphicsUtilities.h"
+#include "tote_bag/gui/lookandfeel/LookAndFeel.h"
+#include "tote_bag/gui/utilities/GraphicsUtilities.h"
 
 #include <juce_gui_basics/juce_gui_basics.h>
