+++ libs/tote_bag/dsp/EnvelopeDetector.cpp
