+++ libs/tote_bag/dsp/EnvelopeDetector.h
