+++ libs/tote_bag/dsp/Saturation.cpp
@@ -9,7 +9,7 @@
 */
 
 #include "Saturation.h"
-#include "Common/DSP/AudioHelpers.h"
+#include "tote_bag/dsp/AudioHelpers.h"
 
 Saturation::Saturation (Type sType, double asymm)
     : saturationType (sType)
