+++ libs/tote_bag/gui/components/panels/VerticalMeterPanel.h
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "Common/GUI/LookAndFeel/LookAndFeel.h"
+#include "tote_bag/gui/lookandfeel/LookAndFeel.h"
 
 #include <juce_gui_basics/juce_gui_basics.h>
