+++ libs/tote_bag/gui/components/widgets/FlatTextButton.cpp
@@ -9,7 +9,7 @@
 */
 
 #include "FlatTextButton.h"
-#include "Common/GUI/LookAndFeel/LookAndFeel.h"
+#include "tote_bag/gui/lookandfeel/LookAndFeel.h"
 
 namespace tote_bag
 {
