+++ libs/tote_bag/gui/components/widgets/LabelSlider.cpp
