+++ libs/tote_bag/gui/components/widgets/LabelSlider.h
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "Common/GUI/Components/Widgets/ParameterSlider.h"
+#include "tote_bag/gui/components/widgets/ParameterSlider.h"
 
 #include <juce_gui_basics/juce_gui_basics.h>
