+++ libs/tote_bag/gui/components/widgets/ParameterSlider.cpp
