+++ libs/tote_bag/gui/lookandfeel/MeterLookAndFeelMethods.h
