+++ libs/tote_bag/gui/managers/ToteBagPresetManager.cpp
