+++ libs/tote_bag/gui/managers/ToteBagPresetManager.h
