+++ libs/tote_bag/gui/utilities/GraphicsUtilities.cpp
