+++ libs/tote_bag/gui/utilities/GraphicsUtilities.h
